+++ crates/subspace-block-relay/src/lib.rs
@@ -69,22 +69,6 @@ pub trait RelayServer {
     async fn on_request(&mut self, request: IncomingRequest);
 }
 
-/// The pool backend used by the protocol client/server sides.
-pub trait PoolBackend<DownloadUnitId, ProtocolUnitId>
-where
-    DownloadUnitId: Encode + Decode,
-    ProtocolUnitId: Encode + Decode,
-{
-    /// Returns all the protocol units for the given download unit.
-    fn download_unit_members(
-        &self,
-        id: &DownloadUnitId,
-    ) -> Result<Vec<(ProtocolUnitId, Vec<u8>)>, RelayError>;
-
-    /// Returns the protocol unit contents with the given Id.
-    fn protocol_unit(&self, id: &ProtocolUnitId) -> Option<Vec<u8>>;
-}
-
 /// Errors returned by the server side.
 #[derive(Debug, Encode, Decode)]
 pub enum RelayError {
